requirements.txt
# Base requirements
Flask==3.0.2
Flask-Login==0.6.3
PyMySQL==1.1.0
Werkzeug==3.0.1
mysql-connector-python==8.3.0
PyJWT==2.8.0
python-dotenv==1.0.1

# Computer Vision and ML
opencv-python==4.9.0.80
numpy==1.26.4
torch==2.2.1
ultralytics==8.2.2

# WebRTC and async
aiortc==1.10.1
aiohttp==3.9.3
quart==0.19.4
quart-cors==0.7.0
quart-auth==0.9.1
aiomysql==0.2.0
hypercorn==0.16.0